--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    ["es2015", { "modules": false }]
-  ],
-  "plugins": [
-    "transform-object-rest-spread"
-  ],
-  "env": {
-    "test": {
-      "presets": ["es2015"]
-    }
-  }
-}
