+++ .gitignore
@@ -1,11 +1,13 @@
 # builds
 /dist
+/markdown/dist
 
 # editors
 /.idea
 
 # dependencies
 /node_modules
+/markdown/node_modules
 
 # misc.
 npm-debug.log*
