+++ markdown/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "prosemirror-suggestions-markdown",
+  "version": "0.2.0",
+  "description": "Markdown additions for prosemirror-suggestions",
+  "main": "dist/index.js",
+  "module": "src/index.js",
+  "author": "Tristan Pemble",
+  "license": "Apache-2.0",
+  "peerDependencies": {
+    "prosemirror-markdown": "^1.1.1",
+    "prosemirror-suggestions": "^0.2.0"
+  },
+  "devDependencies": {
+    "rollup": "^0.63.5",
+    "rollup-plugin-buble": "^0.19.2"
+  },
+  "scripts": {
+    "build": "rollup -c",
+    "watch": "rollup -c -w",
+    "prepare": "npm run build",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  }
+}
