+++ package.json
@@ -1,31 +1,22 @@
 {
-  "name": "@quartzy/prosemirror-suggestions",
-  "version": "0.1.1",
+  "name": "prosemirror-suggestions",
+  "version": "0.2.0",
   "description": "ProseMirror plugin for suggestions (i.e. mentions, tags)",
   "main": "dist/index.js",
+  "module": "src/index.js",
   "author": "Tristan Pemble",
   "license": "Apache-2.0",
-  "peerDependencies": {
-    "prosemirror-markdown": "^0.24.0",
-    "prosemirror-state": "^0.24.0",
-    "prosemirror-view": "^0.24.0"
+  "dependencies": {
+    "prosemirror-state": "^1.2.2",
+    "prosemirror-view": "^1.4.2"
   },
   "devDependencies": {
-    "@types/jest": "^21.1.2",
-    "babel": "^6.23.0",
-    "babel-core": "^6.26.0",
-    "babel-jest": "^21.2.0",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "jest": "^21.2.1",
-    "prosemirror-markdown": "^0.24.0",
-    "prosemirror-schema-basic": "^0.24.0",
-    "prosemirror-state": "^0.24.0",
-    "prosemirror-view": "^0.24.0",
-    "regenerator-runtime": "^0.11.0",
-    "rollup": "^0.50.0",
-    "rollup-plugin-babel": "^3.0.2"
+    "@types/jest": "^23.3.1",
+    "jest": "^23.4.2",
+    "prosemirror-schema-basic": "^1.0.0",
+    "rollup": "^0.63.5",
+    "rollup-plugin-buble": "^0.19.2",
+    "npm-check-updates": "^2.14.2"
   },
   "scripts": {
     "build": "rollup -c",
