+++ src/index.js
@@ -2,9 +2,5 @@ export { suggestionsPlugin, triggerCharacter } from './suggestions';
 export { addTagNodes, tagNodeSpec } from './tags';
 export {
   addMentionNodes,
-  mentionNodeSpec,
-  markdownParser,
-  markdownSerializer,
-  addMentionsToMarkdownParser,
-  addMentionsToMarkdownSerializer,
+  mentionNodeSpec
 } from './mentions';
